[TEST] Write a .travis.yml file that simplifies the test
The modeling test will be split into two tests, one for the gaussian reflector and the other for the multilayer model.
