+++ .travis.yml
@@ -1,13 +1,12 @@
 branches:
- only:
-  - master
-  - /^develop.*$/
-
+  only:
+    - master
+    - /^develop.*$/
+    - teste
 jobs:
-  include:
-    - stage: modeling
-      env: TEST_SUITE=test_gaussianModelBuild
-      - env: TEST_SUITE=test_multiLayerModelBuild
-      script: ./test_travis_docker.sh $TEST_SUITE
-    - stage: deploy
-      script: echo "TODO: Deploy"
+ include:
+  - stage: modeling
+    script: ./test_travis_docker.sh test_gaussianModelBuild
+  - script: ./test_travis_docker.sh test_multiLayerModelBuild
+  - stage: deploy
+    script: echo "TODO deploy"
