+++ experiments/modelingAndPEFInterpolation/gaussianModel/SConstruct
@@ -47,7 +47,7 @@ pef('dataCube',
 	dm=0.025,
 	nt=1001,
 	dt=0.004,
-	nhi=21,
+	nhi=3,
 	a1=a1,
 	a2=a2,
 	rect1=rect1,
@@ -61,7 +61,7 @@ pef('interpolatedDataCube',
 	dm=0.0125,
 	nt=1001,
 	dt=0.004,
-	nhi=21,
+	nhi=3,
 	a1=a1,
 	a2=a2,
 	rect1=rect1,
