+++ experiments/modelingAndPEFInterpolation/multiLayerModel/SConstruct
@@ -71,7 +71,7 @@ pef(dataCube='multiLayerDataCube',
     dm=0.025,
     nt=1001,
     dt=0.004,
-    nhi=21,
+    nhi=3,
     a1=a1,
     a2=a2,
     rect1=rect1,
@@ -85,7 +85,7 @@ pef('interpolatedDataCube',
     dm=0.0125,
     nt=1001,
     dt=0.004,
-    nhi=21,
+    nhi=3,
     a1=a1,
     a2=a2,
     rect1=rect1,
